+++ jax/interpreters/xla.py
@@ -14,7 +14,6 @@
 
 
 from collections import namedtuple, defaultdict
-from distutils.util import strtobool
 import itertools as it
 import operator as op
 import os
@@ -22,7 +21,7 @@ import os
 from absl import logging
 import numpy as onp
 
-from ..config import flags
+from ..config import flags, bool_env
 from .. import core
 from .. import ad_util
 from .. import tree_util
@@ -44,10 +43,10 @@ from . import masking
 
 FLAGS = flags.FLAGS
 flags.DEFINE_bool('jax_debug_nans',
-                  strtobool(os.getenv('JAX_DEBUG_NANS', "False")),
+                  bool_env('JAX_DEBUG_NANS', False),
                   'Add nan checks to every operation.')
 flags.DEFINE_bool('jax_log_compiles',
-                  strtobool(os.getenv('JAX_LOG_COMPILES', "False")),
+                  bool_env('JAX_LOG_COMPILES', False),
                   'Print a message each time a `jit` computation is compiled.')
 
 def _map(f, *xs): return tuple(map(f, *xs))
@@ -171,14 +170,23 @@ def xla_primitive_callable(prim, *arg_specs, **params):
     handlers = tuple(map(partial(aval_to_result_handler, device), aval_out))
     handle_result = lambda xs: tuple(h(x) for h, x in zip(handlers, xs.destructure()))
   tuple_args = len(avals) > 100
-  built_c = primitive_computation(prim, backend, tuple_args, *avals, **params)
+  if prim in initial_style_translations:
+    nreps = initial_style_primitive_replicas(params)
+  else:
+    nreps = 1
+  built_c = primitive_computation(prim, AxisEnv(nreps), backend, tuple_args,
+                                  *avals, **params)
   options = xb.get_compile_options(
       num_replicas=1,
       num_partitions=1,
       device_assignment=device and (device.id,))
   compiled = built_c.Compile(compile_options=options, backend=backend)
-  return partial(_execute_compiled_primitive, prim, compiled, backend,
-                 tuple_args, handle_result)
+  if nreps == 1:
+    return partial(_execute_compiled_primitive, prim, compiled, backend,
+                  tuple_args, handle_result)
+  else:
+    return partial(_execute_replicated_primitive, prim, compiled, backend,
+                   tuple_args, handle_result)
 
 def _device_from_arg_devices(devices):
   """Given devices of inputs, determine where to perform a computation.
@@ -199,7 +207,7 @@ def _device_from_arg_devices(devices):
     raise ValueError(msg.format(", ".join(names)))
 
 @cache()
-def primitive_computation(prim, backend, tuple_args, *avals, **params):
+def primitive_computation(prim, axis_env, backend, tuple_args, *avals, **params):
   c = xb.make_computation_builder("primitive_computation_{}".format(prim.name))
   c.SetOpMetadata(xc.OpMetadata(
       op_type=prim.name,
@@ -209,13 +217,13 @@ def primitive_computation(prim, backend, tuple_args, *avals, **params):
   # return val always set as a side-effect on c
   if prim in backend_specific_translations[platform]:
     rule = backend_specific_translations[platform][prim]
-    rule(c, *xla_args, **params)  
+    rule(c, *xla_args, **params)
   elif prim in translations:
     rule = translations[prim]
     rule(c, *xla_args, **params)
   elif prim in initial_style_translations:
     rule = initial_style_translations[prim]
-    rule(c, AxisEnv(), extend_name_stack(prim.name), *xla_args, backend=backend, **params)
+    rule(c, axis_env, extend_name_stack(prim.name), *xla_args, backend=backend, **params)
   else:
     raise NotImplementedError("XLA translation rule for {} not found".format(prim))
   c.ClearOpMetadata()
@@ -228,7 +236,7 @@ def primitive_computation(prim, backend, tuple_args, *avals, **params):
     raise RuntimeError(msg)
 
 def primitive_subcomputation(prim, *avals, **params):
-  return primitive_computation(prim, None, False, *avals, **params)
+  return primitive_computation(prim, AxisEnv(1), None, False, *avals, **params)
 
 def _execute_compiled_primitive(prim, compiled, backend, tuple_args,
                                 result_handler, *args):
@@ -239,6 +247,17 @@ def _execute_compiled_primitive(prim, compiled, backend, tuple_args,
   out_buf = compiled.Execute(input_bufs)
   if FLAGS.jax_debug_nans:
     check_nans(prim, out_buf.destructure() if prim.multiple_results else out_buf)
+  return result_handler(out_buf)
+
+def _execute_replicated_primitive(prim, compiled, backend, tuple_args,
+                                  result_handler, *args):
+  input_bufs = [
+      [device_put(x, device) for x in args if x is not token]
+      for device in compiled.local_devices()]
+  if tuple_args:
+    input_bufs = [[make_tuple(bufs, device, backend)] for bufs, device in
+                  zip(input_bufs, compiled.local_devices())]
+  out_buf = compiled.ExecutePerReplica(input_bufs)[0]
   return result_handler(out_buf)
 
 def check_nans(prim, bufs):
@@ -268,9 +287,8 @@ def jaxpr_literals(jaxpr):
   return it.chain.from_iterable(eqn_literals(eqn) for eqn in jaxpr.eqns)
 
 def eqn_literals(eqn):
-  if eqn.bound_subjaxprs:
-    (subjaxpr, _), = eqn.bound_subjaxprs
-    for literal in jaxpr_literals(subjaxpr):
+  if eqn.bound_subjaxpr:
+    for literal in jaxpr_literals(eqn.bound_subjaxpr):
       yield literal
   if eqn.primitive in initial_style_translations:
     for param in eqn.params.values():
@@ -322,10 +340,8 @@ def jaxpr_subcomp(c, jaxpr, backend, axis_env, consts, name_stack, *args):
       ans = rule(c, *in_nodes, replica_groups=replica_groups, **new_params)
     elif eqn.primitive in call_translations:
       new_params = check_backend_params(eqn.params, backend)
-      (subjaxpr, const_bindings), = eqn.bound_subjaxprs
-      const_nodes = _map(read, const_bindings)
       rule = call_translations[eqn.primitive]
-      ans = rule(c, subjaxpr, axis_env, const_nodes, in_nodes,
+      ans = rule(c, eqn.bound_subjaxpr, axis_env, in_nodes,
                  name_stack, backend=backend, **new_params)
     else:
       msg = "XLA translation rule for primitive '{}' not found"
@@ -354,14 +370,16 @@ def check_backend_params(params, outer_backend):
 
 
 class AxisEnv(object):
-  def __init__(self, nreps=1, names=None, sizes=None, devices=None):
+  def __init__(self, nreps, names=(), sizes=(), devices=None):
+    assert isinstance(names, tuple)
+    assert isinstance(sizes, tuple)
     self.nreps = nreps
-    self.names = names if names else []
-    self.sizes = sizes if sizes else []
+    self.names = names
+    self.sizes = sizes
     self.devices = devices
 
 def extend_axis_env(env, name, size):
-  return AxisEnv(env.nreps, env.names + [name], env.sizes + [size], env.devices)
+  return AxisEnv(env.nreps, env.names + (name,), env.sizes + (size,), env.devices)
 
 def axis_read(axis_env, axis_name):
   return max(i for i, name in enumerate(axis_env.names) if name == axis_name)
@@ -386,17 +404,21 @@ def _axis_groups(nrep, mesh_spec, mesh_axes):
 def jaxpr_replicas(jaxpr):
   return max(it.chain([1], (eqn_replicas(eqn) for eqn in jaxpr.eqns)))
 
+# TODO(mattjj): this function assumes that only pmap has a parameter named
+# axis_size, and that it corresponds to cross-replica mapping
 def eqn_replicas(eqn):
-  if eqn.bound_subjaxprs:
-    (subjaxpr, _), = eqn.bound_subjaxprs
-    return eqn.params.get('axis_size', 1) * jaxpr_replicas(subjaxpr)
+  if eqn.bound_subjaxpr:
+    return eqn.params.get('axis_size', 1) * jaxpr_replicas(eqn.bound_subjaxpr)
   elif eqn.primitive in initial_style_translations:
-    nums = (jaxpr_replicas(param if type(param) is core.Jaxpr else param.jaxpr)
-            for param in eqn.params.values()
-            if type(param) in (core.Jaxpr, core.TypedJaxpr))
-    return max(it.chain([1], nums))
+    return initial_style_primitive_replicas(eqn.params)
   else:
     return 1
+
+def initial_style_primitive_replicas(params):
+  nums = (jaxpr_replicas(param if type(param) is core.Jaxpr else param.jaxpr)
+          for param in params.values()
+          if type(param) in (core.Jaxpr, core.TypedJaxpr))
+  return max(it.chain([1], nums))
 
 # TODO(mattjj,skyewm): the functions here are utilities for checking if
 # not-yet-supported features are used with multi-host programming
@@ -405,9 +427,8 @@ def jaxpr_has_pmap(jaxpr):
   return any(eqn_has_pmap(eqn) for eqn in jaxpr.eqns)
 
 def eqn_has_pmap(eqn):
-  if eqn.bound_subjaxprs:
-    (subjaxpr, _), = eqn.bound_subjaxprs
-    return jaxpr_has_pmap(subjaxpr)
+  if eqn.bound_subjaxpr:
+    return jaxpr_has_pmap(eqn.bound_subjaxpr)
   elif eqn.primitive in initial_style_translations:
     return any(jaxpr_has_pmap(param if type(param) is core.Jaxpr else param.jaxpr)
                for param in eqn.params.values()
@@ -420,9 +441,8 @@ def jaxpr_collectives(jaxpr):
   return it.chain.from_iterable(eqn_collectives(eqn) for eqn in jaxpr.eqns)
 
 def eqn_collectives(eqn):
-  if eqn.bound_subjaxprs:
-    (subjaxpr, _), = eqn.bound_subjaxprs
-    for c in jaxpr_collectives(subjaxpr):
+  if eqn.bound_subjaxpr:
+    for c in jaxpr_collectives(eqn.bound_subjaxpr):
       yield c
   elif eqn.primitive in initial_style_translations:
     for param in eqn.params.values():
@@ -492,7 +512,7 @@ def _xla_callable(fun, device, backend, name, *arg_specs):
   xla_consts = _map(c.Constant, consts)
   xla_args = _xla_callable_args(c, abstract_args, tuple_args)
   out_nodes = jaxpr_subcomp(
-      c, jaxpr, backend, AxisEnv(nreps, [], []), xla_consts,
+      c, jaxpr, backend, AxisEnv(nreps, (), ()), xla_consts,
       extend_name_stack(wrap_name(name, 'jit')), *xla_args)
   built = c.Build(c.Tuple(*out_nodes))
 
@@ -596,16 +616,15 @@ xla_call = partial(core.call_bind, xla_call_p)
 xla_call_p.def_custom_bind(xla_call)
 xla_call_p.def_impl(_xla_call_impl)
 
-def _xla_call_translation_rule(c, jaxpr, axis_env, const_nodes,
+def _xla_call_translation_rule(c, jaxpr, axis_env,
                                in_nodes, name_stack, backend, name, device=None):
   del device  # Ignored.
   subc = xb.make_computation_builder("jit_{}".format(name))
-  consts = [subc.ParameterWithShape(c.GetShape(n)) for n in const_nodes]
   args = [subc.ParameterWithShape(c.GetShape(n)) for n in in_nodes]
-  out_nodes = jaxpr_subcomp(subc, jaxpr, backend, axis_env, consts,
+  out_nodes = jaxpr_subcomp(subc, jaxpr, backend, axis_env, (),
                             extend_name_stack(name_stack, wrap_name(name, 'jit')), *args)
   subc = subc.Build(subc.Tuple(*out_nodes))
-  return c.Call(subc, list(const_nodes) + list(in_nodes))
+  return c.Call(subc, list(in_nodes))
 ad.primitive_transposes[xla_call_p] = partial(ad.call_transpose, xla_call_p)
 
 
@@ -643,10 +662,11 @@ def lower_fun(fun, instantiate=False, initial_style=False):
   """Build a translation rule for a traceable function."""
   def f(c, *args, **params):
     backend = params.pop('backend', None)
+    # TODO(mattjj): revise this 'calling convention'
     if initial_style:
       axis_env, name_stack, xla_args = args[0], args[1], args[2:]
     else:
-      axis_env, name_stack, xla_args = AxisEnv(), '', args
+      axis_env, name_stack, xla_args = AxisEnv(1), '', args
     xla_shapes = tuple(map(c.GetShape, xla_args))
     avals = map(_aval_from_xla_shape, xla_shapes)
     pvals = [pe.PartialVal((a, core.unit)) for a in avals]
@@ -962,19 +982,18 @@ ad.deflinear(device_put_p, lambda cotangent, **kwargs: [cotangent])
 masking.defvectorized(device_put_p)
 
 
-def _remat_translation_rule(c, jaxpr, axis_env, const_nodes, in_nodes,
+def _remat_translation_rule(c, jaxpr, axis_env, in_nodes,
                             name_stack, backend, name, device=None, concrete=None):
   # This looks a lot like _xla_call_translation_rule, except for a widget we use
   # to foil CSE.
   del device, concrete  # Unused.
   subc = xb.make_computation_builder("remat_call_subcomputation")
-  consts = [subc.ParameterWithShape(c.GetShape(n)) for n in const_nodes]
   args = [subc.ParameterWithShape(c.GetShape(n)) for n in in_nodes]
   args = [_foil_cse(subc, x) for x in args]
-  out_nodes = jaxpr_subcomp(subc, jaxpr, backend, axis_env, consts,
+  out_nodes = jaxpr_subcomp(subc, jaxpr, backend, axis_env, (),
                             extend_name_stack(name_stack, wrap_name(name, 'remat')), *args)
   subc = subc.Build(subc.Tuple(*out_nodes))
-  return c.Call(subc, list(const_nodes) + list(in_nodes))
+  return c.Call(subc, list(in_nodes))
 call_translations[pe.remat_call_p] = _remat_translation_rule
 
 def _foil_cse(c, x):